Add Kubernetes plugin integration tests
