+++ test/integration/k8scontroller/helpers.go
@@ -116,6 +116,8 @@ func RunSlothController(ctx context.Context, config Config, ns string, cmdArgs s
 		fmt.Sprintf("SLOTH_KUBE_CONTEXT=%s", config.KubeContext),
 		fmt.Sprintf("SLOTH_KUBE_NAMESPACE=%s", ns),
 		fmt.Sprintf("SLOTH_DEVELOPMENT=%t", true),
+		fmt.Sprintf("SLOTH_DEVELOPMENT=%t", true),
+		fmt.Sprintf("SLOTH_SLI_PLUGINS_PATH=%s", "./"),
 	}
 
 	return RunSloth(ctx, env, config.Binary, fmt.Sprintf("kubernetes-controller %s", cmdArgs), true)
